+++ packages/indexer/src/api/endpoints/tokens/get-tokens/v5.ts
@@ -28,7 +28,7 @@ export const getTokensV5Options: RouteOptions = {
   description: "Tokens",
   notes:
     "Get a list of tokens with full metadata. This is useful for showing a single token page, or scenarios that require more metadata.",
-  tags: ["api", "Tokens"],
+  tags: ["api", "x-deprecated"],
   plugins: {
     "hapi-swagger": {
       order: 9,
@@ -1154,7 +1154,10 @@ export const getTokensV5Options: RouteOptions = {
                   feeBreakdown: feeBreakdown,
                 }
               : undefined,
-            royaltiesPaid: query.includeRoyaltiesPaid ? r.royalties_paid : undefined,
+            royaltiesPaid:
+              query.includeRoyaltiesPaid && r.royalties_paid !== null
+                ? r.royalties_paid
+                : undefined,
           },
         };
       });