+++ packages/indexer/src/models/collections/index.ts
@@ -13,7 +13,7 @@ import {
 import { Tokens } from "@/models/tokens";
 import MetadataApi from "@/utils/metadata-api";
 import * as royalties from "@/utils/royalties";
-import * as marketplaceFees from "@/utils/marketplace_fees";
+import * as marketplaceFees from "@/utils/marketplace-fees";
 import { logger } from "@/common/logger";
 
 export class Collections {
@@ -213,7 +213,7 @@ export class Collections {
           orders.source_id_int,
           orders.valid_between
         FROM tokens
-        JOIN orders
+        LEFT JOIN orders
         ON tokens.floor_sell_id = orders.id
         WHERE tokens.collection_id = $/collection/
         ORDER BY tokens.floor_sell_value