+++ packages/indexer/src/websockets/opensea/index.ts
@@ -10,7 +10,6 @@ import {
 import { ItemListedEventPayload, ItemMetadataUpdatePayload } from "@opensea/stream-js/dist/types";
 import * as Sdk from "@reservoir0x/sdk";
 import { WebSocket } from "ws";
-
 import { logger } from "@/common/logger";
 import { redis } from "@/common/redis";
 import { now } from "@/common/utils";
@@ -18,10 +17,8 @@ import { config } from "@/config/index";
 import { OpenseaWebsocketEvents } from "@/models/opensea-websocket-events";
 import { PartialOrderComponents } from "@/orderbook/orders/seaport";
 import { generateHash, getSupportedChainName } from "@/websockets/opensea/utils";
-
 import * as orderbookOrders from "@/jobs/orderbook/orders-queue";
 import * as orderbookOpenseaListings from "@/jobs/orderbook/opensea-listings-queue";
-
 import { handleEvent as handleItemListedEvent } from "@/websockets/opensea/handlers/item_listed";
 import { handleEvent as handleItemReceivedBidEvent } from "@/websockets/opensea/handlers/item_received_bid";
 import { handleEvent as handleCollectionOfferEvent } from "@/websockets/opensea/handlers/collection_offer";